electronapp/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
sounddevice==0.4.6
numpy==1.24.3
